Refactor cl_nav2z components and client behaviors into header-only implementations.
