+++ smacc2_client_library/cl_nav2z/include/cl_nav2z/client_behaviors/cb_nav2z_client_behavior_base.hpp
@@ -30,7 +30,7 @@ namespace cl_nav2z
 class CbNav2ZClientBehaviorBase : public smacc2::SmaccAsyncClientBehavior
 {
 public:
-  virtual ~CbNav2ZClientBehaviorBase();
+  virtual ~CbNav2ZClientBehaviorBase() {}
 
   template <typename TOrthogonal, typename TSourceObject>
   void onStateOrthogonalAllocation()
