+++ smacc2_client_library/cl_nav2z/include/cl_nav2z/client_behaviors/cb_pause_slam.hpp
@@ -28,8 +28,32 @@ namespace cl_nav2z
 class CbPauseSlam : public smacc2::client_behaviors::CbServiceCall<slam_toolbox::srv::Pause>
 {
 public:
-  CbPauseSlam(std::string serviceName = "/slam_toolbox/pause_new_measurements");
-  void onEntry() override;
+  CbPauseSlam(std::string serviceName = "/slam_toolbox/pause_new_measurements")
+  : smacc2::client_behaviors::CbServiceCall<slam_toolbox::srv::Pause>(serviceName.c_str())
+  {
+  }
+
+  inline void onEntry() override
+  {
+    this->requiresComponent(this->slam_);
+
+    auto currentState = slam_->getState();
+
+    if (currentState == CpSlamToolbox::SlamToolboxState::Resumed)
+    {
+      RCLCPP_INFO(
+        getLogger(), "[CbPauseSlam] calling pause service to toggle from resumed to paused");
+      this->request_ = std::make_shared<slam_toolbox::srv::Pause::Request>();
+      CbServiceCall<slam_toolbox::srv::Pause>::onEntry();
+      this->slam_->toggleState();
+    }
+    else
+    {
+      this->request_ = nullptr;
+      RCLCPP_INFO(
+        getLogger(), "[CbPauseSlam] calling skipped. The current slam state is already paused.");
+    }
+  }
 
 protected:
   CpSlamToolbox * slam_;
