+++ smacc2_client_library/cl_nav2z/include/cl_nav2z/components/waypoints_navigator/cp_waypoints_event_dispatcher.hpp
@@ -1586,7 +1586,11 @@ public:
   template <typename TDerived, typename TOrthogonal>
   void initialize(ClNav2Z * client);
 
-  void postWaypointEvent(int index);
+  inline void postWaypointEvent(int index)
+  {
+    auto & fn = postWaypointFn[index % WAYPOINTS_EVENTCOUNT];
+    if (fn != nullptr) fn();
+  }
 };
 
 template <typename TEv>
